Fix flapping update process
The m_ok variable was not initialized, and set to true in case of
successful initialization - it depended on pure luck.

Since debug versions usually initialize stack and heap allocated memory
(to 0xCD or something like that), this explains why update check was
failing silently and randomly in release versions.

Signed-off-by: Simon Rozman <simon@rozman.si>